Struct massa_channel::receiver::MassaReceiver

pub struct MassaReceiver<T> {
    pub(crate) receiver: Receiver<T>,
    pub(crate) name: String,
    pub(crate) actual_len: Gauge,
    pub(crate) received: Counter,
    pub(crate) ref_counter: Arc<()>,
}

Fields
receiver: Receiver<T>name: Stringactual_len: Gauge

channel size

received: Counter

total received messages

ref_counter: Arc<()>

reference counter to know how many receiver are cloned

Implementations

impl<T> MassaReceiver<T>

pub fn update_metrics(&self)

increment manually the metrics Should be used when using the receiver with select! macro select! does not call recv()

fn unregister_metrics(&self)

unregister metrics

pub fn try_recv(&self) -> Result<T, TryRecvError>

attempt to receive a message from the channel

pub fn recv_deadline(&self, deadline: Instant) -> Result<T, RecvTimeoutError>

pub fn recv_timeout(&self, timeout: Duration) -> Result<T, RecvTimeoutError>

pub fn recv(&self) -> Result<T, RecvError>